Added Spring Data JPA implementation services

+++ pet-clinic-data/src/main/java/com/liferon/petclinic/model/Visit.java
@@ -12,7 +12,7 @@ import java.time.LocalDate;
 @Entity
 @Table(name = "visits")
 @EqualsAndHashCode(exclude = {"pet"})
-public class Visit extends BaseEntity<Integer> {
+public class Visit extends BaseEntity<Long> {
 
     private LocalDate date;
     private String description;
